requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
mysqlclient==2.1.1
sqlalchemy==2.0.15
passlib==1.7.4
python-multipart==0.0.6
PyJWT==2.10.1
PyMySQL==1.1.1